shapescape_texture_tools/requirements.txt
Pillow~=9.0
better-json-tools>=1.0.3,~=1.0
